FastAPI/TestSiteForSandingMoney/requirements.txt
SQLAlchemy~=2.0.43
fastapi~=0.116.1
pydantic~=2.11.7
starlette~=0.47.3
aiosmtplib~=4.0.2
python-dotenv~=1.1.1
passlib~=1.7.4